Dernière fonction ajoutée (reste à revoir les détails)

+++ api-mafermelocale/app/Http/Controllers/API/UserController.php
@@ -4,10 +4,11 @@ namespace App\Http\Controllers\API;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
+use App\Http\Resources\FarmCollection;
 use Illuminate\Support\Facades\Auth;
-use Validator;
 use App\Models\User;
 use App\Http\Resources\UserResource;
+use App\Models\Farm;
 use Spatie\QueryBuilder\QueryBuilder;
 
 class UserController extends BaseController
@@ -48,6 +49,17 @@ class UserController extends BaseController
         return $this->sendResponse(new UserResource($user), 'User retrieved');
     }
 
+    public function showFarm($id, Farm $farm)
+    {
+        $farm = Farm::find($id);
+
+        if (is_null($farm)) {
+            return $this->sendError('The farm does not exist.');
+        }
+
+        return $this->sendResponse(FarmCollection::collection($farm), 'Farm retrieved');
+    }
+
     /**
      * Update the specified user with the requested parameters
      * 
@@ -55,20 +67,19 @@ class UserController extends BaseController
      */
     public function update(Request $request, $id)
     {
-
         $user = User::find($id);
 
         if (is_null($user)) {
             return $this->sendError('The user does not exist.');
         }
 
-        if (Auth::guard('sanctum_user')->user()->id != $id) {
-            return $this->sendError('You can\'t modify another user !', [], 403);
+        if (Auth::guard('sanctum_user')->user()->id != $id) { // If the user is not the same as the authenticated user (the one who is trying to update the user)
+            return $this->sendError('You can\'t modify another user !', [], 403); // Forbidden !
         }
 
-        $user->update($request->all());
+        $user->update($request->all()); // Update the user with the requested parameters
 
-        return $this->sendResponse(new UserResource($user), 'User updated.');
+        return $this->sendResponse(new UserResource($user), 'User updated.'); // Return the updated user
     }
 
     /**
@@ -78,14 +89,18 @@ class UserController extends BaseController
      */
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
+        $user = User::find($id);
 
-        if (Auth::guard('sanctum_user')->user()->id != $id) {
-            return $this->sendError('You can\'t delete another user !', [], 403);
+        if (is_null($user)) { 
+            return $this->sendError('The user does not exist.'); // Return error if user not found 
+        }
+ 
+        if (Auth::guard('sanctum_user')->user()->id != $id) { // If the user is not the same as the authenticated user (the one who is trying to delete the user)
+            return $this->sendError('You can\'t delete another user !', [], 403); // Forbidden !
         }
 
-        $user->delete();
+        $user->delete(); // Delete the user
 
-        return $this->sendResponse([], 'User deleted.');
+        return $this->sendResponse([], 'User deleted.'); // Return a success message that the user has been deleted
     }
 }
